requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.12.0
Flask==3.1.2
google-auth==2.40.3
google-genai==1.36.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
numpy==2.3.3
openai==1.107.3
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
